Move `EquipmentBroker` to its own file.
